requirements.txt
groq
python-dotenv
langchain
langchain[groq]
streamlit
langchain_community
psycopg2
langchain_huggingface
unstructured[md]
smolagents
pandas
mcp
mcp[cli]
smolagents[openai]
smolagents[mcp]
matplotlib